Split S7739 rule into rule/decorator/false-positives/helpers
The FP-remediation logic for the no-thenable exceptions had grown to
554 lines inline in rule.ts. Split following the S6747 convention:
rule.ts now only wires the upstream rule through decorate(); decorator.ts
holds the interceptReport wiring; false-positives/index.ts holds the
thenable-specific exception detection; helpers.ts holds the generic,
non-FP-specific AST utilities.
